we can now deploy sites without having repository connected

+++ app/Services/Site/SiteService.php
@@ -19,6 +19,7 @@ use App\Services\DeploymentServices\Ruby\Ruby;
 use App\Contracts\Server\ServerServiceContract as ServerService;
 use App\Contracts\RemoteTaskServiceContract as RemoteTaskService;
 use App\Contracts\Repository\RepositoryServiceContract as RepositoryService;
+use function response;
 
 class SiteService implements SiteServiceContract
 {
@@ -119,7 +120,10 @@ class SiteService implements SiteServiceContract
      */
     public function deploy(Server $server, Site $site, SiteServerDeployment $siteServerDeployment, SiteDeployment $oldSiteDeployment = null)
     {
-        $this->repositoryService->importSshKey($site);
+        if($site->userRepositoryProvider) {
+            $this->repositoryService->importSshKey($site);
+        }
+
         $deploymentService = $this->getDeploymentService($server, $site, $oldSiteDeployment);
 
         foreach ($siteServerDeployment->events as $event) {
@@ -185,19 +189,27 @@ class SiteService implements SiteServiceContract
 
     /**
      * @param Site $site
-     * @return Site $site
+     * @return Site|\Illuminate\Http\JsonResponse
      */
     public function createDeployHook(Site $site)
     {
+        if(!$site->userRepositoryProvider) {
+            return response()->json('The site does not have a connected repository', 400);
+        }
+
         return $this->repositoryService->createDeployHook($site);
     }
 
     /**
      * @param Site $site
-     * @return Site $site
+     * @return Site|\Illuminate\Http\JsonResponse
      */
     public function deleteDeployHook(Site $site)
     {
+        if(!$site->userRepositoryProvider) {
+            return response()->json('The site does not have a connected repository', 400);
+        }
+
         return $this->repositoryService->deleteDeployHook($site);
     }
